bug7515: fixed build under Debian 7
hg-changeset: 942c3b3a0177bce4ce68c5860e2f0d02603faf42

+++ cimple/lib/Time.cpp
@@ -64,7 +64,7 @@ void Time::sleep(uint64 timeout_usec)
 
 #else
 
-    timeval tv = { timeout_usec / 1000000, timeout_usec % 1000000 };
+    timeval tv = { long(timeout_usec / 1000000), long(timeout_usec % 1000000) };
     select(0, 0, 0, 0, &tv);
 
 #endif
